Major updates! (#44)
* added No Patching checkbox to extractor widget

* updated models adapters to support no_patching

* updated segmentation widget

* removed unused imports

* fixed some typings

* added the dataset class

* init using the new dataset for feature extraction

* init using zarr storage

* moved up no patching checkbox

* updated extractor widget and no_patching option

* updated segmentation widget using zarr storage

* clean up & formatting

* fixed extraction progress info

* updated dataset to handle images type inside

* updated prediction pipeline over a large stack

* fixed run pipeline; file dialogs open in parent dialog

* updated run_pipeline script

* removed unused import in run_pipeline

* fixed models params: image height & width as int

* ignored some typing

* updated requirements

* fixed image height & width as int

* fixed dataset & get_model_ready_image image dimention problem

* added dataset test

* fixed adapter: concat double output into one tensor

* updated mobilesam test

* updated dino adapter test

* added sam2 adapter test

* added pipeline_prediction test

* bumped version

* run_pipeline script can be used to only extract features

* updated dependencies in pyproject.toml

* fixed bug for calculating padding with no_patching

* revert back storage to hdf5; improved get_train_data

* fixed embedding_extraction test

+++ tests/model_adapter_tests/test_dino_adapter.py
@@ -2,10 +2,10 @@ import numpy as np
 import pytest
 import torch
 import torch.nn as nn
+from embedding_extraction import check_embedding_extraction
 
 from featureforest.models.DinoV2 import DinoV2Adapter, get_model
 from featureforest.utils.data import get_stack_dims
-from embedding_extraction import check_embedding_extraction
 
 
 class MockDinoEncoder(nn.Module):
@@ -22,7 +22,7 @@ class MockDinoEncoder(nn.Module):
         return output, None
 
 
-def get_mock_model(img_height: float, img_width: float) -> DinoV2Adapter:
+def get_mock_model(img_height: int, img_width: int) -> DinoV2Adapter:
     model = MockDinoEncoder()
     device = torch.device("cpu")
     dino_model_adapter = DinoV2Adapter(model, img_height, img_width, device)
@@ -49,10 +49,10 @@ def test_mock_adapter(test_patch: np.ndarray):
 
     result_real = real_adapter.model.get_intermediate_layers(
         transformed_input_patch_real, 1, return_class_token=False, reshape=True
-    )[0]
+    )[0]  # type: ignore
     mock_result = mock_adapter.model.get_intermediate_layers(
         transformed_input_patch_mock
-    )[0]
+    )[0]  # type: ignore
 
     assert len(result_real) == len(mock_result)
     assert result_real[0].shape == mock_result[0].shape
